Refactor using rich.
